+++ packages/skift.hjert.kernel/includes/kernel/dev/pci.h
@@ -1,7 +1,8 @@
 #pragma once
 
-/* Copyright © 2018-2019 MAKER.                                               */
-/* This code is licensed under the MIT License.                               */
-/* See: LICENSE.md                                                            */
-
 #include <skift/generic.h>
+
+typedef struct {
+} pci_dev_t;
+
+// vim:set ts=4 sw=4 et:
