--- packages/skift.hjert.kernel/includes/kernel/pci.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <skift/generic.h>
-
-typedef struct {
-} pci_dev_t;
-
-// vim:set ts=4 sw=4 et:
